Fix #391: Show number of items in file list
Show the number of files/recordings and the number of directories
(excluding '..' and 'Trash') in the selection list in the space
above the list.

Add the volume size to the volume usage information below the
selection list.

[DiskInfo]
Implement SIZE and USED functions.
Extend the size multipliers to use TB for values larger than 10 TB.
Use 10 ** x notation for large powers-of-10 numbers.
Factor out size conversion to string to a separate method.
Allow the item label to be specified when a DiskInfo object is created.
Some PEP8 fixes.

[MovieList]
Add calculation of the number of directories and the number of files.
Change the UI text "Trashcan" to "Trash".

[MovieSelection]
Add diskSize, numFiles and numFolders widget implementations and
code to update them.

[Trashcan]
Extend the size multipliers to use TB for values larger than 10 TB.
Use 10 ** x notation for large powers-of-10 numbers.
Change the UI text "Trashcan" to "Trash".
Some PEP8 fixes.

Requires corresponding changes in branch Implement-enhancement-391
in repository skin-default.