Rework interface to activating Timeshift
Add new general method InfoBarTimeshift.activateTimeshift() that should be easier to extend than the old InfoBarTimeshift.activateTimeshiftEnd(). New specific methods InfoBarTimeshift.activateTimeshiftPause(), InfoBarTimeshift.activateTimeshiftSkipBackLeft() and InfoBarTimeshift.activateTimeshiftSkipBackDown() for use in the ActionMap. Retained the old methods InfoBarTimeshift.activateTimeshiftEnd() and InfoBarTimeshift.activateTimeshiftEndAndPause() as calls on the new InfoBarTimeshift.activateTimeshift() for backward compatibility, but marked them as deprecated.

This is mainly to make the implementation of issue #382 cleaner, but the old logic was a bit messy anyway.