Show infobar & time shift progress bar when switching between timeshift events
Show the infobar when sswitching between timeshift events with
PREV/NEXT and "Show infobar on skip forward/backward" is enabled.

The infobar and time shift progress bar will also be shown if the
switch in time shift event is simply a consequence of playing past
the end of a time shift event if "Show infobar on event change" is
enabled.