Fix bug #398: Crash when pressing "TEXT" on Network services setup numeric settings fields
[NetworkSetup]
Remove VirtualKeyboardActions context and showVirtualKeyboard actions
from ActionMaps in the setup screens. These actions weren't being
disabled when the focus was on a ConfigNumber entry (in the DLNA
and DYN-DNS setup screens), and KeyText() was being called on the
ConfigNumber entries. ConfigNumber is nt compatible with KeyText().

Removed from all ActionMaps, even for screens with no ConfigNumber
entries to prevent future problems if a ConfigNumber entry was
added.