More VirtualKeyboardActions cleanup [CCcamInfo, CronTimer, Setup, TimerEntry]
Remove incorrectly used or redundant instances of context
VirtualKeyboardActions and action showVirtualKeyboard.

Where these were in an assignment to self["VirtualKB"], they were
redundant, but innocuous. Where they were in a different action
map, they had the potential to cause a recurrence of bug 398 if a
ConfigNumber item was added to the config screen that they implemented.