Simplify and generalise testing for debug logging
Simplify the scripts and reduce the amount of re-testing in checking
whether debug logging is enabled and in extracting the log file
directory.

Disentangled the debug path generation from running enigma2 and
moved to early in the code so that more parts of the script are
able to log to the debug log if necessary.

Put @libpath@ into a variable at the start of the script (makes
testing outside the build environment simpler).

Reordered the enigma exit code comments into numerical order.

I haven't been able to test within the build environment.