[MovieSelection] Fix "movie_sort" screen control, improve sort button prompts, optimise language translation strings and improve menu generation code.
67c67
< 	(str(MovieList.SORT_GROUPWISE), _("Recordings by date, other media by name"), 'New->Old & A->Z'),
---
> 	(str(MovieList.SORT_GROUPWISE), _("Recordings by date, other media by name"), 'Rec New->Old & A->Z'),

This change allows for the on screen prompt for two different sorts to be made noticeably different.

71c71
< 	(str(MovieList.SORT_ALPHA_DATE_NEWEST_FIRST_FLAT), _("Flat by name, then by date"), 'A->Z, New->Old Flat'),
---
> 	(str(MovieList.SORT_ALPHA_DATE_NEWEST_FIRST_FLAT), _("Flat by name, then by date"), 'Flat A->Z, New->Old'),

This change makes the long and short descriptions more easily comparable.  The word "Flat" was lost at the end of the line.  Further the longer description has the word "Flat" at the start of the text so it would be clearer and more uniform if the button description matched the long description more closely.

75c75
< 	(str(MovieList.SORT_ALPHAREV_DATE_OLDEST_FIRST_FLAT), _("Flat by reverse name, then by reverse date"), 'Z->A, Old->New Flat'),
---
> 	(str(MovieList.SORT_ALPHAREV_DATE_OLDEST_FIRST_FLAT), _("Flat by reverse name, then by reverse date"), 'Flat Z->A, Old->New'),

Same as previous.

331c331
< 		MultiContentEntryText(pos=(20, 8), size=(575, 50), font=0, text=_(name)),
---
> 		MultiContentEntryText(pos=(20, 0), size=(575, 50), font=0, flags=RT_HALIGN_LEFT | RT_VALIGN_CENTER, text=_(name)),

This change makes the on screen menu formatting more automatically adaptable and immune to font and/or itemHeight changes.  The change requires no changes or adjustment to the existing skin.

361c361
< 		menu.append(MovieMenuEntryComponent((_("Sort by") + "..."), csel.selectSortby))
---
> 		menu.append(MovieMenuEntryComponent((_("Sort") + "..."), csel.selectSortby))

This change makes the "Sort" option prompts better match the equivalent "On end of movie" options.  That is, "Sort" makes the change without the pop up menu while "Sort..." brings up a selection menu to select the sort order. The is functionally more mnemonic.  This is the same operational behaviour and conforms to the "On end of movie" functionality.

Also, for the multi-language interface the multiple versions of the "Sort" and "On end of movie" strings should be unified and the punctuation (ellipsis) separated.  For example, _("Sort") + "..." and _("On end of movie") + "..." so that each version only needs one translation entry for the multiple use strings.

642c642
< 				'sortby': (_("Sort order"), _("Change sort order...")),
---
> 				'sortby': (_("Sort" + "..."), _("Select sort order from menu")),

Reworked to better match the more mnemonic "On end of movie" prompts.

646c646
< 				'movieoff_menu': (_("On end of movie..."), _("Select end-of-movie action from menu")),
---
> 				'movieoff_menu': (_("On end of movie") + "..."), _("Select end-of-movie action from menu")),

This change is for multi-language translation optimisation.

1353,1356d1352
< 		if index == MovieList.SORT_ALPHA_DATE_OLDEST_FIRST:
< 			index = MovieList.SORT_ALPHANUMERIC
< 		elif index == MovieList.SORT_ALPHAREV_DATE_NEWEST_FIRST:
< 			index = MovieList.SORT_ALPHANUMERIC_REVERSE

This code is removed to allow all 10 sort orders to be individually addressable within the skin.  All the sort order are now different so mapping any sort onto another icon is inappropriate.  The code as it currently exists inhibits the skin's "movie_sort" variable from representing *all* the appropriate sort states.