Change factory reset to be run as an "after" task in enigma2.sh
Add factory reset as an "after" task in enigma2.sh run on exit code
40 from enigma2.

Modify [GeneralSetup, Standby] to exit with exit code 40 instead
of implementing the factory reset internally and exiting with exit
code 2 to trigger a reboot.

Add data/factoryReset/50-enigma2 to be run out of the enigma2.sh
"after" task in enigma2.sh using run-parts. 50-enigma2 should be
installed in /usr/share/enigma2/factoryReset on the PVR.

The 50-enigma2 script emulates the previous internal factory reset
actions, with the exception that the "reset" version of
/etc/network/interfaces is a "here" document in the script.

This should be replaced by a file copy when a more general mechanism
for restoring files to their factory reset state is decided on.

Add factoryReset to SUBDIRS in data/Makefile.am.

I haven't been able to test within the build environment. The build
environment needs to ensure that the PVR image is built with any
run-parts scripts in /usr/share/enigma2/factoryReset have execute
permission on (as in the repository).