The CutList Editor Plugin appears to have a keymap error / limitation.
The internal keymap file "/usr/lib/enigma2/python/Plugins/Extensions/CutListEditor/keymap.xml" only enables the colour buttons for the "dreambox remote control (native)".  This means that the colour coded playback control buttons are NOT mapped to the colour buttons on the Beyonwiz.

Also the STOP and MENU buttons are unmapped.

This commit makes the colour buttons always available and maps MENU to the menu function and STOP to the leave function.