Allow for national decoding of program classifications
Rename EventName class to EventNameBasic and extend it as EventName
with code to allow the definition of program classifications
(Rating/SmallRating) that are specific to the rating's country code.

If a table isn't provided, fall back to the ETSI EN 300 468 decoding.

Provide classification tables for Australian classifications.

Also allow Country=countrycode argument to Rating and SmallRating to
override the broadcast country code from the skin.

Add "editing" arguments for text display; Trimmed/NotTrimmed to strip
(not strip) whitespace from the start/end of text from the EPG, and
Separated/NotSeparated to add (not add) a blank line between the
description and extended description in FullDescription, NextDescription
and ThirdDescription. Defaults to NotTrimmed,Separated.