Fix Bug #399: [TimerEntry] SMS-style text entry help stays on screen when on-screen virtual keyboard is used
[TimerEntry]
Implement hideTextHelp() and showTextHelp as a calls to HideHelp()
and ShowHelp() in ConfigListScreen.

Add ShowHelp and HideHelp to onExecBegin and onExecEnd respectively
to hide and show the SMS-style text entry help as required.

Removed unnecessary calls to hideTextHelp() and showTextHelp() in
showHelp(), KeyText() and renameEntryCallback().

Changes affect all text entry fields in [TimerEntry].