Small correction on cuesheet support servicemp3.
I noticed that if there was no cue file avbl and user did not insert a cut,
an empty cuts file was made on exit media.
With this patch there will be no empty cuts file made for nothing.
If there was a cuts file and the user removed all the cuts, the file will be removed as well.

	modified:   lib/service/servicemp3.cpp