Fix bug #405: Delete in Movie Player can stop/delete the wrong recording
The problem was caused because if two recordings had the Name in the timer,
they would be named as "<timecode> <recname>" and "<timecode> <recname>_1".

If rec_filename was "<timecode> <recname>", and there were recording
files "<timecode> <recname>.ts" and "<timecode> <recname>_1.ts"
with the same <timecode> and <recname>, the he test in
MovieSelection.delete() "rec_filename in timer.Filename" would
succeed for both files.

When MENU>Setup>TV>Recording settings>Composition of recording
filenames is set to "Short filenames", <timecode> is just the date,
so this problem is much more likely to occur with that setting.