Workaround bug #408: On T4, when recordings start simultaneously, one may be empty
The problem seemed to have been triggered when multiple tuners were
being accessed concurrently, so this code forces sequential access
to the tuners in eDVBFrontend::tuneLoopInt(), which seems to work
around the problem.

Forcing sequential access to the tuners increases the tuner config
time from < 1 sec for concurrent access to ~2-3 sec on a T4 for
sequential access.

This code has logging on the queuing code that forces sequential
access that can probably be removed for production code. If the
changes are to be accepted, I can strip out the logging and tidy
up the branch.

This workaround does not address the underlying cause of the problem,
which is probably in T4-specific hardware or in the T4 driver code.