Fix bug #403: T4 OLED Screen Cleanup during HDMI-IN operation
The default OLED picon needs to be greyscale without alpha plane.

/usr/share/enigma2/lcd_picon_default.png was greyscale with alpha
plane.

Alpha plane removed, scaling tweaked and filled to full-screen
170x40 pixels.

This allows the default "Beyonwiz button" picon to be displayed on
the OLED screen if there is no specific picon for a channel.

This part of the fix is not HDMI IN-specific.