Fix bug #403: T4 OLED Screen Cleanup during HDMI-IN operation
Picons for HDMI IN don't display on either the main screen or the
OLED screen, because eServiceHDMI::getInfoString(sServiceref) returns
an empty string.

Added return of the serviceref string to
eServiceHDMI::getInfoString(int w).

Display of the HDMI IN picons also requires branch Add-hdmi-in
in the picons-australia repository.

This problem is HDMI IN-specific, but not OLED display-specific.