Revert "picload: refactor filling surface data"
This reverts commit 9e2dc00a0179abe9229f42e482250a487c768bc9.

Since picload is used almost exclusively for showing pictures in the photo
viewer and similar plugins, it usually handles images that have roughly
the target size of the framebuffer, up to 1920x1080.
Just filling this with zeroes takes about the same amount of time as the
actual data copying. Filling the whole frame roughly halves performance.

The code isn't optimized for readability, but for speed and with good
reason.