eServiceReference cleanup - 2 rewrite addLcnToList & writeBouquet
Rewrite addLcnToList() to have it comply with FreeTV OP 41 "Logical
Channel Descriptor and Allocation of Logical Channel Numbers", Table
2 overlap cases 2 and 3.

Overlap case 1 seems to be dealt with in the C++ scan code (in
particular eDVBScan::addKnownGoodChannel()).

Assign duplicate LCNs sequentially from LCN 350 instead of at
orig_lcn + 350, which goes outside the range for LCN duplicates
(350..399) if orig_lcn >= 50.

Drop weaker duplicates if keeping them would require the duplicate
to be assigned an LCN > 399.

The method addLcnToList() was reamed to addLcn() as a function
nested in LCN.read().

Also simplify it and make it more efficient.

Rewrite writeBouquet() to simplify it. Also add a marker before
"extra" services are added to the list (like HDMI In).

Change the test for inserting markers so that the "Duplicates"
marker appears in the correct place if a duplicate is assigned to
LCN 300 (as can now happen with the new addLcn() code).

Duplicate handling has only been tested on simulated data.