Fix bug #411: Colons in folder names can cause confusion - rework
Fix error in commit 5fc6609 that meant that when returning to a
parent directory in the MoviePlayer, the initial focus wasn't on
the previous directory.

There were possibly other bugs with caused by the same error.

Fix eServiceReference constructors for file system directories that
were in the form:
    ref = eServiceReference(
        eServiceReference.idFile,
        eServiceReference.isDirectory,
        path
    )

and should have been:
    ref = eServiceReference(
        eServiceReference.idFile,
        eServiceReference.noFlags,
        1
    )
    ref.setPath(path)

Symbolic names are needed for the reference path type for filesystem
servicerefs.

Also use eServiceReference.noFlags in some constructors instead of
0.