Fix skin_path_prefix in [Pixmap,MultiPixmap]
Replace os.path with None as the default when fetching the screen
skin_path attribute in MultiPixmap. If a "~/" was in a pixmaps
attribute path, the old code would still throw an exception in
Tools.Directories.resolveFilename(), but it was not the right
exception. The new code will throw an AssertionError, as intended
in these circumstances.

The optional path_prefix arg to resolveFilename() is set explicitly
to None, even though that is its default, to emphasise that there
is no fallback path available if the skin_path attribute is not set
in the widget's screen.

Removed unused fetch of the screen skin_path attribute in Pixmap.