picexit: do not maintain jpeg section data
Eacht jpgeg section is used at most once so there
is no need to keep all the section in memory. It is
sufficient to just process the last read section.
Also no need to maintain the length values in the
data buffer as it is already saved in the itemlen
variable.