[InfoBar] Save resume point on end of file, instead of deleting it.
This change effectively enables the ability to determine if a file has
been watched or not. The old code would remove the resume point, thus
marking the file as unwatched. The new code leaves the resume point at
the end of the file, thus indicating that the file has been watched to
the end.