[InfobarGenerics] Remove redundant code for checking stale resume point paths.
The mount point path check was redundant. findMountPoint() will always
return a path that will evaluate ismount() as True, because that is the
test it uses to find the mount point.

Document the problems with this function. Path lookups on network mounts
could be potentially very expensive. Some kind of an asynchronous mechanism
is needed to clean up the resume point cache.