Servref cleanup 3 - redo serviceref types
The DVB serviceref types enum was originally mistakenly put in
ServiceID, which is for the DVB SID, not for the serviceref type.

Moved the DVB serviceref types from ServiceID to eServiceReferenceDVB.

Also created eServiceReferenceFS mainly as a placeholder for a
service reference types enum.

eServiceReferenceFS service reference types enum will be used in a
following commit.