Fix bug #429: In recording settings default location settings, LEFT & RIGHT do nothing
Copy the ConfigList variables for the default recording locations
into the corresponding config.usage variables in
RecordingSettings.changedEntry(), so that the ConfigList variables
don't get over-written by the old values of the config.usage variables
when RecordingSettings.createSetup() is called.

Also tidy up all the calls to self["config"].getCurrent() after
it is assigned to self.item with the use of self.item and an
"a == str1 or a == str2" ... with "a in (str1, str2, ...)".