Fix bug #442: Buttons 1/3 don't change the time scale correctly in Graphical EPG
[EpgSelection] In EpgSelection.keyNumberGlobal():

Save changes to config.epgselection.graph_prevtimeperiod as the expected
string, rather than as an integer, so that the changes are accepted by the
config variable,

Use .value = instead of .setValue() when changing
config.epgselection.graph_prevtimeperiod.

Modify the tests on timeperiod so that pressing 1/3 doesn't take
timeperiod out of the allowed range of
config.epgselection.graph_prevtimeperiod.