Fix bug #445: Time navigation issues in Graphical EPG
[EpgList]

Replace the combined use of self.time_base and self.offs to control
time navigation (as self.time_base + self.offs * self.time_epoch * 60)
with the use of only self.time_base. This avoids the problems in
the bug report where navigation can get stuck.

Replace attribute self.offs with a property with a set function
that does nothing and a get function that always returns 0, for backward
compatibility of a sort.

Retain self.resetOffset() as a method that does nothing.

[EpgSelection]

Remove references to self['list'].resetOffset().