Fold common code together in EpgSelection keyNumberGlobal()
[EpgSelection]

In self.keyNumberGlobal(), fold together the code for the number
key operations in the graphical EPG and the InfoBar EPG by
parameterising them with the appropriate config variables.

This fix also has the side-effect of fixing the equivalent to bug
#443 in the InfoBar EPG. The InfoBar EPG instance of the bug was
missed when bug #443 was fixed for the graphical EPG.

Also simplify the expression in the calculation of the basetime
tuple in the '9' (go to prime time) button press by using tuple
slices.