Fix bug #454: Using live TV panic button can cause a crash
[InfoBarNumberZap]

In doPanicButton() use a valid formulation for the root bouquet
when Enable multiple bouquets is disabled.

Implement the action for doPanicButton() when Enable multiple
bouquets is disabled.  Use the current root bouquet for the search
for the panic button service instead of the fixed Terrestrial TV
LCN that was used in the original code, so that the use isn't
suddenly taken to a different bouquet.

Extract common service search code for the two states of Enable
multiple bouquets into a new method.

Use more up-to-date initialisation of the root bouquet when Enable
multiple bouquets is enabled.

Prevent doPanicButton() going into an infinite loop when Enable
multiple bouquets is enaabled, but the root serviceref contains no
bouquet servicerefs, or when no bouquet contains any servicerefs.

Add more testing for error conditions.