RecordTimer.py: Store custom flags in timer entry
Add a "flags" field to RecordTimerEntry. This contains a set of items
that plugins can store into the timer. Each entry should be a string.

In particular, a plugin like AutoTimer could place the word "autotimer"
in this field to remember who created it, instead of hacking the object
and misusing the tags field for this.