Fix issue #302-2: Combine end time and duration popups for instant record menu
The code allows for buttons to be assigned to increment/decrement
the hours and minutes of the Time and Duration fields. However,
consensus counld not be reached about which buttons to use, so the
functions have bee disabled by commenting out the relevant code.

In keymap.xml:

Add a context to support increment/decrement keys in TimeDateDurationInput.
Currently commented out.

In config.py:

[config]

Add definitions for KEY_PAGEUP, KEY_PAGEDOWN keys for new/modified
classes.

[ConfigDateTime]

Allow for a second increment value, base time, callback on
select/deslelect of the config item and enable/disable constraint
that the value is >= base time.

Allow objects to use functions other than localtime() & mktime()
for conversions between Unix time and a time_struct.

Add KEY_PREV, KEY_NEXT as keys that increment/decrement
using the second increment value.

Add controllable limiting of the value to be >= tthe base time.

Allow formatstring to be a callable that returns a conversion of
the time as well as to be a string

Add methods for conversion between the object's value and
value - base time.

[ClockTime]

Mixin class to allow a ConfigDateTime object to display as and
interact with the user as a ConfigClockTime. The class is designed
to make use of the base time in its containing ConfigDateTime

[ConfigClockTime]

ConfigDateTime with a ClockTime mixin to allow time entry.

[ConfigDuration]

New class derived from ConfigDateTime that displays and
interacts with value - base time as a duration. Normally
prevents durations < 0, but displays negative durations if
negative durations (i.e. value < time base) if the limiting of value
is disabled.

[ConfigClockDuration]

ConfigDuration with a ClockTime mixin to allow time entry.

In TimeDateInput.py:

[TimeDateInputBase]

Abstract common parts of TimeDateInput and new class TimeDateDurationInput
into TimeDateInputBase.

Reverse the order of ConfigListScreen and Screen so that ConfigListScreen
methods over-ride Screen methods. In particular

Add symbolic names for the "save" masks.

[TimeDateInput]

In TimeDateInput.createConfig() change the incorrect creation of
the default conf_time as a tuple to be just the plain object (remove
trailing comma).

[TimeDateDurationInput]

New class extending TimeDateInput to allow the definition of a
date/time/duration screen.

Also adds the ability to put descriptive text into the Config items inside
a TimeDateDurationInput object.

In order to properly handle changing the base time limit on the
config items, TimeDateDurationInput.getTimestamp() returns a Config
object instead of an integer to capture changes to the Config item's
value after getTimestamp() is called.

In InfoBarGenerics.py:

[InfoBarInstantRecord]

Remove redundant "or False" from "(logical expression) or False" construct.

Use new TimeDateDurationInput instead of TimeDateInput to enter end
time or duration instead of only end time for "Record until given
time" and "Change instant recording end time". The corresponding
"duration" menu entries remain with their previous UI, and will be
removed in a later "menu simplification" checkin.

In InfoBarInstantRecord.setEndTime() add keyword argument "new" to
distinguish whether the call is to set the length of a recording
or to modify it. This is currently only used to modify the description
of the Duration config item, but it will be used in future checkins
to change the behaviour of the popup on exit: to continue recording
with the original time if it was modifying the end time (instead
of stopping the timer, and to delete the recording as well as the
timer if the call was to set the end time.

In InfoBarInstantRecord.TimeDateInputClosed() adapt to the new interface to
TimeDateDurationInput.close()

Dependencies:

Requires modifications in the corresponding branches in the
easy-skin-aus-hd and skin-full-metal-wizard repositories.