[ipk.py] Fix .ipk show contents tar errors
This change detects if the .ipk file was created by the ar archiver.  If so, the archive will be unpacked by ar rather that using tar and failing.  If the ar test fails and the file was not created by ar then fall back to the original code and try opening the file / archive with tar.