lib/base/eerror.h: Allow compiler to optimize things away Move the level check into the header so the compiler can skip what it doesn't need. For example:
eLog(3, 0, "%d", a*b/c);

When the log level is set to 4, the "a*b/c" expression will not be evaluated.

This also allows for a much simpler way of removing debug levels from code. If
the MAX level is lower than the log level, the whole statement will be removed
by the compiler, since it can "see" that the code is unreachable. But it will
still detect syntax errors.

Unfortunately, this will never work for string tags.