Fix issue #302-2: warn user if repeated instant record actions are made
[InfoBarGenerics]

Add code to check whether a recording is already running on the
current service when an instant recording is started, and ask what
the user wants to do.

Link into the new code in Timeshift to check whether the current
or a prior timeshift event has already been saved.

[Timeshift]

Add code to check whether a prior timeshift event has already been
saved if it is saved again, and ask what the user wants to do.

Add code to check whether the timeshift buffer for the program
currently being broadcast is already being marked to be saved and
ask what the user wants to do.