Fix issue #302-2: clean up recording if Set recording duration is cancelled
[InfoBarGenerics]

If "Create instant recording..." is selected, but the time setting
popup is cancelled, the timer will be stopped and the short recording
that was created while the popup was displayed will be deleted.

Unqueue "A recording has been started ..." popup when time setting
is cancelled, so that the user doesn't get the popup if the recording
is cancelled & deleted.

[RecordTimer]

Add a unique id to the "A recording has been started ..." popup
so that it can be cancelled or overridden.