Fix issue #303: Single-press record
[InfoBarGenerics]

Make instant record actions table-driven so that the tables for the
instant record menu can be shared with the setup for binding instant
record actions to long- and short-REC button presses.

Allow two extra button actions that aren't in the menu for true
single-button actions to record for duration and seve the currently
viewed timeshift buffer.

[UsageConfig], setup.xml

Add entries to configure the button presses.

keymap.xml

Allow long- and short-REC to be distinguished from
short-RED.