Fix bug #381: GREEN button text overlaid in EPG Search after Add Timer
Add Components.Label.DummySource that allows a text widget to appear
in a <widget source= ... /> element in a screen skin but not display
on the screen.

Use Components.Label.DummySource Components.Button.Button and
Components.Label.Label so that when they are used in
<widget source= ... /> widgets, the apparent intention of the code
to make the widget inactive and not display, is completed, so that
neither text arguments to the constructor nor text= widget attributes
will be displayed on the screen.

This prevents text overwriting when the same Component instance
appears in both a <widget source= ... /> and a <widget name= />
skin widget.

Also change GUIComponent.applySkin() to Label.applySkin() in
MultiColorLabel.applySkin(), which it should have been anyway, but
which is now required by the changes to Label.