Fix bug #381: GREEN button text overlaid in EPG Search after Add Timer - refix
The original fix for this bug set the Label/Button text to ""
too early, so that any references to widget.getText() in the
widget's Screen.__init__() returned "".

This caused problems in other parts of the UI, including crashing
when number buttons were used to change channel in live TV.

This fix sets the widget text field at the normal time, but
saves it and sets the text to "" when a downstream
Converter/Renderer connects, so that the downstream receives
"" when it reads widget.text in the connection process.

The text is restored to its correct value in widget.applySkin(),
as in the original fix.

This should prevent any user code "seeing" the changed text value
as "".