Fix issue #302-3: Fixed fallback duration for "padded event" instant recording
If the user chooses "padded event" for the default duration of an
instant recording and no event information is available, make the
default recording duration 120 min instead of a 24-hour "indefinite"
recording.

Also modify the "recording started" popup so that it shows the
fallback time instead of "padded event" if the fallback time is
used.