Fix bug #468: Incorrect "unknown action" messages from ChoiceBox ActionMaps
Change the priority of self["actions"] so that it is searched before
self["cancelaction"]. This eliminates most of the spurious warnings.

Add a "drop through" mapping for "back" in self["actions"] so
that it no longer produces a spurious warning.

Remove unused contexts from self["cancelaction"] and change it from
NumberActionMap to ActionMap. This is just tidying-up and doesn't itself
change the warnings once the other changes have been made.