Fix issue #456: Add UI support for Discrete IR Codes
[keymap.xml]

In GlobalActions, distinguish front panel POWER make and break
actions from remore POWER make and break actions so that short-POWER
from the remote can be handled differently from short-POWER on the
front panel.

In GlobalActions and StandbyActions, add new mappings for POWER2,
SUSPEND and WAKEUP.

[setup.xml]

Add an "expert" level setting to enable/disable short-POWER actions
from the remote.

Re-order the setting entries for the settings for short- and
long-POWER and short- and long-REC actions so that the entry for
short-POWER immediately follows the new setting that disables it
from the remote.

Use two entries for the short-POWER actions with "requires" attributes
to change the setting label and description to reflect changes in
the enable/disable short-POWER actions setting.

[UsageConfig]

Add two new config variables, config.usage.short_power_enable and
config.usage.short_power_disable. The config.usage.short_power_disable
config variable is marked NoSave, and is tied to alwas be the
negation of config.usage.short_power_enable. These support the text
switching mentioned above in keymap.xml, and are required because
"requires" attributes can't contain expressions.

[Standby]

Add a powerKey global that is set in mytest.runScreenTest() to
reference the same object as mytest.power. This is required by
Standby.deepStandby().

Add Standby.deepStandby() to allow the POWER2 key to shut down the
PVR from standby, and call it on the "deepstandby" action.

Move code to enable and disable video output into methods
Standby.videoOn() and Standby.videoOff(). They are used in
PowerKey.shutdown() to allow an information popup to be displayed.

[mytest]

Add methods PowerKey.conditional_powerdown() and
PowerKey.conditional_powerdown() to only act on short-POWER
from the remote if short-POWER actions are enabled for the remote.
The functions return 0 if short-POWER actions are disabled for the remote
so that they can fall through to the POWER actions for the front panel.

Add an action on "deepstandby" to implement shutdown on POWER2.

If the PVR is recording in standby when shut down, show a different
message from shutdown while recording when running ("Remaining in
standby ..." instead of "Entering standby...") and switch on video
output so that the message can be seen on the TV screen if the TV
is on. Turn off video again when te message box closes.

Set Screens.Standby.powerKey = mytest.power do that StandbydeepStandby()
can access PowerKey.shutdown().

Import RecordTimer and use the RecordTimer.AFTEREVENT.DEEPSTANDBY
symbol instead of a fixed integer.