Fix bug #473: Repeat power timers not repeating
The begin/end times for timers were being reset to the "orig" value
inappropriately when PowerTimerEntry.timeChanged() was called at
the end of TimerEntry.processRepeated().

Used PowerTimerEntry.processRepeated() wrapper to correctly handle
the begin/end times so that they are reset to "orig" before calling
TimerEntry.processRepeated(), flagged as not to be used by
PowerTimerEntry.timeChanged(), and then set new "orig" values when
TimerEntry.processRepeated() returns.