Fix #477: Long-CH+/- changes both PiP and main screen channel, not just PiP
[keymap.xml]

Use long/break codes for CH+/- instead of make/repeat to match the
expectation of the code.

[InfoBarGenerics]

Add a toggle on long-press of CH+/- so that if the long press doesn't
result in a new screen (like the "Please open Picture in Picture
first" error popup) the "break" action for the button press doesn't
get carried out.