Fix #476: Odd FAV behaviour after zap timer changes bouquet
[ChannelSelection]

Ensure that ChannelSelection.instance is initialised only when a
ChannelSelection object is initialised, and not when a derived
object is created (like a PiPZapSelection).

Correct the assert text to refer to ChannelSelection, not InfoBar.

[InfoBarChannelSelection]

Remove the unnecessary use of ChannelSelection.instance in zapUp()
and zapDown() and replace it with self.servicelist2.

The original code depended on ChannelSelection.instance being
incorrectly initialised to the single system instance of PiPZapSelection
(which is also referenced by self.servicelist2).