[RecordTimerEntry] code cleanup in _zapToTimerService()
In _zapToTimerService():

Abstract the check of whether a serviceref is in a bouquet into a function,

Recode the loop over the bouquets so that it doesn't an loop endlessly
if the zap serviceref isn't in any bouquet.

Add log messages for when the zap fails.

This step doesn't affect Bug #475, but it stops the hang in Bug #474.
However, if MENU>Setup>TV>Channel selection>Enable multiple bouquets
is disabled, the zap will still fail, but now with an error message
in the log.