Fix Bug #474: Zap timer causes UI hang when multiple bouquets disabled
[RecordTimerEntry]

In _zapToTimerService():

Move the search through the bouquets into the "true" side of the
if test on config.usage.multibouquet, and add code to search the
Favourites list only in the "false" side of the test.

This means that "zap" and "zap and record" timers now both no longer
hang the UI (previous commit) and function correctly (this commit).

Only do a ChannelSelectionInstance.enterPath() on one bouquet when
config.usage.multibouquet is disabled, so that the title on the
channel list is displayed correctyl. Move the
ChannelSelectionInstance.clearPath() to immediately before the
enterPath() calls to bring all the path list operations together.

Use a predefined base eServiceRef to define the multi-bouquet root
serviceref. This must have been missed when the serviceref cleanup
was done.