Fix Bug #475: A zap timer can unnecessarily change the selected bouquet
[RecordTimerEntry]

If config.usage.multibouquet is enabled, scan the bouquet list
twice, the first time to check whether the zap channel is in the
current bouquet and use that as the zap bouquet (so that the bouquet
doesn't change), then to check whether it is in any other bouquet,
and use the first one found.

This means that if the bouquet will only change if the zap channel
is not in the current bouquet.