Fix Bug #478: No .ap file created for HDMI In recordings - causes Execute cuts to fail
[eEncoder]

Add definitions for fixed PIDs used by the HDMI IN encoder,
and use them in methods to return corresponding the PIDs.

[eServiceHDMIRecord]

In doRecord(), add code to set the video PID for the HDMI IN encoder
as the "timing pid" for the recording, from which the access points
for the .ap file are saved.

Because it's not possible to tell what streams are presend in the
output of the decoder, if the HUDM IN has no video stream,
only audio, no access points will be found.

This has the side-effect of filling and preserving the .sc file
for the recording.