Fix Bug #479: Crash on zap timer start when in media player selection screen
When MEDIA, MENU>Settings>Stop service on return to movie list set to "yes"
an item has been played and stopped, so that the current live TV service isn't
playing in the background, MoviePlayer.instance is not None, but
Navigation.currentlyPlayingServiceReference is None, so when
Navigation.currentlyPlayingServiceReference.toString()is called, the UI crashes.

This fix tests for Navigation.currentlyPlayingServiceReference being None
and avoids calling toString() on it.