Fix Bug #488: "Continue timeshifting" when saving current timeshift cancels timeshift save
[InfoBarTimeshift]

Only set self.save_current_timeshift to False when the action passed to
checkTimeshiftRunningCallback() or saveTimeshiftActions() is "noSave".

Change the "Continue timeshifting" option in the "You have chosen
to save the current timeshift event, but the event has not yet
finished. What do you want to do?" popup to "Continue save timeshift",
because the popup is due to a pending timeshift shave and can happen
when the user is not timeshifted.