Improve description of EPG cache load and save options
The description of the EPG cache load and save functions was not as clear
as it could have been and as a result many users have enabled these
features without understanding what they do. This resulted in cases where
the load option was enabled without a third party source updating the
cache. When coupled with the save option, it was possible to create a
scenario where the system would save a possibly stale version of the EPG
and subsequently would reload this version, thus in the process discarding
any EPG updates that were already received.