Option to stop movie player with exit-button
Option “Allow quit movie player with exit” was extended with new
possible value “Behave like stop-button”. This allows to use
exit-button just like stop-button. In particular when option “Behavior
when a movie is stopped” is set to “Return to movie list” the button
“Exit” returns to movie list too, which was not possible before (it
could be configured to return to previous service but not to movie
list).