Fix Bug #495: "Save timeshift" incorrect on standby when "Timeshift save action on zap" is not "Ask user"
[Standby]

Trigger the close of the (temporary) Standby screen on self.onShow
instead of self.onHide. The latter isn't triggered if the "You have
chosen to save the current timeshift event, but the event has not
yet finished. What do you want to do?" popup isn't shown.

Using self.onShow triggers the close whether the popup is shown or
not.