Fix Bug #494: "Save timeshift" fails on standby when "Show warning when timeshift is stopped" is disabled
[Timeshift]

Change the action of ptsGetTimeshiftStatus() so that it returns
True if self.save_current_timeshift is True, whether
config.usage.check_timeshift.value ("Show warning when timeshift
is stopped") is True or not.

This has the side-effect of raising the "Recording(s) are in progress
or coming up soon!" on shutdown (from POWER or from the Power /
Restart screen) if a timeshift save has been requested even when
"Show warning when timeshift is stopped" is disabled. I think that
is desirable behaviour; it previously simply shut down in those
conditions previously.

[Standby]

Changed the popup text from "Recording(s) are in progress or coming
up soon!" to "Recording(s) are in progress or coming up soon, or
you are saving timeshift!". Also changed the "You seem to be in
timeshift!" popup text to "You seem to be in timeshift or saving
timeshift!".

NB:

When a timeshift save has been requested,
InfoBarTimeshift.checkSavingCurrentTimeshift() incorrectly sets
config.timeshift.isRecording to True. This means that the "Recording(s)
are in progress ..." popup text is used when the "You seem to be
in timeshift ..." text should be displayed. The text for the
"Recording(s) are in progress ..." popup should be "Recording(s)
are in progress or coming up soon, or a timeshift merge is pending!"
if the setting of config.timeshift.isRecording was correct.

The setting of the timeshift save variables will be overhauled in
the fix for Bug #491: Unnecessary delays in performing merge when
saving current event's timeshift.